+++ lib/Constants.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+List kBottomGridText = [0, '00', '.', '=', 1, 2, 3, '+', 4, 5, 6, '-', 7, 8, 9, '*', 'C', '%', '⌫', '/'];
+List kscientificGridText = ['sin', 'cos', 'tan', 'log', 'In', '(', ')', '^', '√', '!', 'π', 'e', 'INV', 'RAD', 'DEG'];
+
+List<Color> ktextColor = [
+  Colors.black,
+  Colors.black,
+  Colors.black,
+  Color(0xFF35d5d2),
+  Colors.black,
+  Colors.black,
+  Colors.black,
+  Color(0xFF35d5d2),
+  Colors.black,
+  Colors.black,
+  Colors.black,
+  Color(0xFF35d5d2),
+  Colors.black,
+  Colors.black,
+  Colors.black,
+  Color(0xFF35d5d2),
+  Colors.black,
+  Colors.black,
+  Colors.black,
+  Color(0xFF35d5d2)
+];
